requirements.txt
Flask==3.0.3
gunicorn==22.0.0
ultralytics==8.3.22
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

torch==2.2.2+cu121
torchvision==0.17.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

yolov5
